+++ Bencodex/Types/Binary.cs
@@ -6,9 +6,7 @@ using System.Diagnostics.Contracts;
 using System.Globalization;
 using System.Linq;
 using System.Runtime.CompilerServices;
-using System.Security.Cryptography;
 using System.Text;
-using Bencodex.Misc;
 
 namespace Bencodex.Types
 {
@@ -53,23 +51,6 @@ namespace Bencodex.Types
         /// <inheritdoc cref="IValue.Kind"/>
         [Pure]
         public ValueKind Kind => ValueKind.Binary;
-
-        /// <inheritdoc cref="IValue.Fingerprint"/>
-        [Pure]
-        public Fingerprint Fingerprint
-        {
-            get
-            {
-                ImmutableArray<byte> digest = _digest is { } cache
-                    ? cache[0] is { } b
-                        ? b
-                        : ByteArray.Length > 20
-                            ? ImmutableArray.Create(SHA1.Create().ComputeHash(ToByteArray()))
-                            : ByteArray
-                    : ImmutableArray<byte>.Empty;
-                return new Fingerprint(Kind, EncodingLength, digest);
-            }
-        }
 
         /// <inheritdoc cref="IValue.EncodingLength"/>
         [Pure]